+++ opentelemetry-sdk/src/metrics/meter.rs
@@ -402,8 +402,6 @@ impl SdkMeter {
             let validation_result = validate_bucket_boundaries(boundaries);
             if let Err(err) = validation_result {
                 // TODO: Include the buckets too in the error message.
-                // TODO: This validation is not done when Views are used to
-                // provide boundaries, and that should be fixed.
                 otel_error!(
                     name: "InstrumentCreationFailed",
                     meter_name = self.scope.name(),